create generic function to create SwerveControllerCommand from any traject
