+++ src/main/java/frc/robot/autos/TestNoZeroGyroAuto.java
@@ -1,7 +1,6 @@
 package frc.robot.autos;
 
 import java.util.List;
-import edu.wpi.first.math.controller.PIDController;
 import edu.wpi.first.math.geometry.Pose2d;
 import edu.wpi.first.math.geometry.Rotation2d;
 import edu.wpi.first.math.trajectory.Trajectory;
@@ -10,6 +9,7 @@ import edu.wpi.first.math.trajectory.TrajectoryGenerator;
 import edu.wpi.first.wpilibj2.command.SwerveControllerCommand;
 import frc.robot.Constants;
 import frc.robot.commands.ZeroMotorsWaitCommand;
+import frc.robot.modules.AutoBase;
 import frc.robot.modules.Vision;
 import frc.robot.subsystems.Swerve;
 
@@ -18,7 +18,6 @@ import frc.robot.subsystems.Swerve;
  */
 public class TestNoZeroGyroAuto extends AutoBase {
     Vision vision;
-    Swerve swerve;
 
     /**
      * Autonomous that aligns limelight then excecutes a trajectory.
@@ -28,7 +27,6 @@ public class TestNoZeroGyroAuto extends AutoBase {
      */
     public TestNoZeroGyroAuto(Swerve swerve, Vision vision) {
         super(swerve);
-        this.swerve = swerve;
         System.out.println("Limelight Auto !!");
         TrajectoryConfig config =
             new TrajectoryConfig(Constants.AutoConstants.kMaxSpeedMetersPerSecond,
@@ -40,11 +38,7 @@ public class TestNoZeroGyroAuto extends AutoBase {
                 new Pose2d(0.1, 0.1, new Rotation2d(0))),
             config);
 
-        SwerveControllerCommand firstHalfTraject = new SwerveControllerCommand(firstHalfTrajectory,
-            swerve::getPose, Constants.Swerve.swerveKinematics,
-            new PIDController(Constants.AutoConstants.kPXController, 0, 0),
-            new PIDController(Constants.AutoConstants.kPYController, 0, 0), thetaController,
-            swerve::setModuleStates, swerve);
+        SwerveControllerCommand firstHalfTraject = baseSwerveCommand(firstHalfTrajectory);
         ZeroMotorsWaitCommand firstWait = new ZeroMotorsWaitCommand(swerve, 3);
 
         addCommands(firstHalfTraject, firstWait);
