+++ src/main/java/frc/robot/modules/AutoBase.java
@@ -1,0 +1,50 @@
+package frc.robot.modules;
+
+import com.pathplanner.lib.PathPlannerTrajectory;
+import com.pathplanner.lib.commands.PPSwerveControllerCommand;
+import edu.wpi.first.math.controller.PIDController;
+import edu.wpi.first.math.controller.ProfiledPIDController;
+import edu.wpi.first.math.trajectory.Trajectory;
+import edu.wpi.first.wpilibj2.command.SequentialCommandGroup;
+import edu.wpi.first.wpilibj2.command.SwerveControllerCommand;
+import frc.robot.Constants;
+import frc.robot.subsystems.Swerve;
+
+/**
+ * Parent class for all autonomous commands
+ */
+public class AutoBase extends SequentialCommandGroup {
+    public Swerve swerve;
+    public ProfiledPIDController thetaController =
+        new ProfiledPIDController(Constants.AutoConstants.kPThetaController, 0, 0,
+            Constants.AutoConstants.kThetaControllerConstraints);
+
+    /**
+     * Autonomous that aligns limelight then excecutes a trajectory.
+     *
+     * @param swerve swerve subsystem
+     */
+    public AutoBase(Swerve swerve) {
+        this.swerve = swerve;
+        thetaController.enableContinuousInput(-Math.PI, Math.PI);
+    }
+
+    public SwerveControllerCommand baseSwerveCommand(Trajectory trajectory) {
+        SwerveControllerCommand command = new SwerveControllerCommand(trajectory, swerve::getPose,
+            Constants.Swerve.swerveKinematics,
+            new PIDController(Constants.AutoConstants.kPXController, 0, 0),
+            new PIDController(Constants.AutoConstants.kPYController, 0, 0), thetaController,
+            swerve::setModuleStates, swerve);
+        return command;
+    }
+
+
+    public PPSwerveControllerCommand baseSwerveCommand(PathPlannerTrajectory trajectory) {
+        PPSwerveControllerCommand command = new PPSwerveControllerCommand(trajectory,
+            swerve::getPose, Constants.Swerve.swerveKinematics,
+            new PIDController(Constants.AutoConstants.kPXController, 0, 0),
+            new PIDController(Constants.AutoConstants.kPYController, 0, 0), thetaController,
+            swerve::setModuleStates, swerve);
+        return command;
+    }
+}
